+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
-- A new pipeline task between `check-temp-table-data` task and `swap-dataset-table` task for transforming data before making public.
+- A new pipeline task between `insert-into-temp-table` task and `check-temp-table-data` task for transforming data before making public.
 - A new field `email_marketing_consent` to Contacts dataset to hold Email consent data joined from Consent dataset.
 
 ## 2020-08-03